+++ Region_mathonly.h
@@ -76,7 +76,7 @@ public:
 	Region_mathonly(const Region_mathonly &src);
 	Region_mathonly& operator=(const Region_mathonly &src);
 
-	Vector3d B(int pointId,const Vector3d &offset); //returns the B field at a given point, allows interpolation between points and offset due to non-ideal beam
+	Vector3d B(size_t pointId,const Vector3d &offset); //returns the B field at a given point, allows interpolation between points and offset due to non-ideal beam
 };
 
 #endif
